+++ lib/neighbor/redis.rb
@@ -8,6 +8,7 @@ require "json"
 require_relative "redis/index"
 require_relative "redis/flat_index"
 require_relative "redis/hnsw_index"
+require_relative "redis/svs_vamana_index"
 require_relative "redis/vector_set"
 require_relative "redis/version"