+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (unreleased)
 
 - Added support for vector sets
+- Added support for SVS Vamana indexes
 - Dropped support for Ruby < 3.2
 
 ## 0.2.1 (2025-05-06)